==> 8.AddingCensusBorders.ipynb <==
***Adding borders to census tract data***

# import Exports/Data/CensusTractExposure.csv
import pandas as pd
import geopandas as gpd

filepath = 'Exports/Data/7.CensusTractExposure.csv'
exposure_data_2010 = pd.read_csv(filepath)

tract_path_2010 = "Imports/CensusTractGIS/2010/US_tract_2010.shp"
gdf_2010 = gpd.read_file(tract_path_2010)

exposure_data_2010 = pd.merge(exposure_data_2010, gdf_2010[['GISJOIN', 'geometry']], on=['GISJOIN'], how='left', indicator=True, validate='1:1')
exposure_data_2010['_merge'].value_counts()

print('Tracts with no border:', exposure_data_2010.query("_merge == 'left_only'")['TRACTCE'].unique())

print('dropping mismatches')
exposure_data_2010 = exposure_data_2010.query("_merge == 'both'")
exposure_data_2010 = exposure_data_2010.drop(columns=['_merge'])

exposure_data_2010 = gpd.GeoDataFrame(exposure_data_2010, geometry='geometry')

print(exposure_data_2010.crs)
exposure_data_2010 = exposure_data_2010.to_crs(epsg=4326)
exposure_data_2010['geometry'] = exposure_data_2010['geometry'].simplify(tolerance=0.001)
print(exposure_data_2010.crs)

exposure_data_2010.head()

exposure_data_2010.to_file('Exports/Data/8.CensusTractExposure.geojson', driver='GeoJSON')

***Doing the same for counties***

county_path_2010 = "Imports/CountyGIS/US_county_2010.shp"
exposure_path_2010 = "Exports/Data/7.CountyExposure.csv"
gdf_2010 = gpd.read_file(county_path_2010)
exposure_data_2010 = pd.read_csv(exposure_path_2010)

gdf_2010.rename(columns={'STATEFP10': 'STATEFP', 'COUNTYFP10': 'COUNTYFP'}, inplace=True)
exposure_data_2010['STATEFP'] = exposure_data_2010['STATEFP'].astype(str)
exposure_data_2010['COUNTYFP'] = exposure_data_2010['COUNTYFP'].astype(str)
exposure_data_2010['GISJOIN'] = 'G' + exposure_data_2010['STATEFP'].str.zfill(2) + exposure_data_2010['COUNTYFP'].str.zfill(4) + '0'

exposure_data_2010 = pd.merge(exposure_data_2010, gdf_2010[['GISJOIN', 'geometry']], on=['GISJOIN'], how='left', indicator=True, validate='1:1')
assert exposure_data_2010['_merge'].unique() == 'both'
exposure_data_2010.drop(columns=['_merge'], inplace=True)

exposure_data_2010 = gpd.GeoDataFrame(exposure_data_2010, geometry='geometry')

print(exposure_data_2010.crs)
exposure_data_2010 = exposure_data_2010.to_crs(epsg=4326)
exposure_data_2010['geometry'] = exposure_data_2010['geometry'].simplify(tolerance=0.001)
print(exposure_data_2010.crs)

exposure_data_2010.head()

# Save as geojson
exposure_data_2010.to_file('Exports/Data/8.CountyExposure.geojson', driver='GeoJSON')
